// src/ts/constant/api.ts
export enum ApiPath {
    GetActuatorAuditevents = '/actuator/auditevents',
    GetActuatorBeans = '/actuator/beans',
    GetActuatorCaches = '/actuator/caches',
    DeleteActuatorCaches = '/actuator/caches',
    GetActuatorHealth = '/actuator/health',
    GetActuatorConditions = '/actuator/conditions',
    GetActuatorConfigprops = '/actuator/configprops',
    GetActuatorEnv = '/actuator/env',
    PostActuatorEnv = '/actuator/env',
    DeleteActuatorEnv = '/actuator/env',
    GetActuatorInfo = '/actuator/info',
    GetActuatorLogfile = '/actuator/logfile',
    GetActuatorLoggers = '/actuator/loggers',
    PostActuatorLoggers = '/actuator/loggers',
    GetActuatorHeapdump = '/actuator/heapdump',
    GetActuatorThreaddump = '/actuator/threaddump',
    GetActuatorMetrics = '/actuator/metrics',
    GetActuatorScheduledtasks = '/actuator/scheduledtasks',
    GetActuatorHttptrace = '/actuator/httptrace',
    GetActuatorMappings = '/actuator/mappings',
    PostActuatorRefresh = '/actuator/refresh',
    GetActuatorFeatures = '/actuator/features',
    GetActuator = '/actuator',
    PostCacheClear = '/cache/clear',
    GetCache = '/cache',
    PostCacheEvict = '/cache/evict',
    GetDiscoveryInstance = '/discovery/instance',
    GetDiscoveryService = '/discovery/service',
    GetToolGenerateRsa = '/tool/generate/rsa',
    GetPingerCheck = '/pinger/check',
    PostPingerCheck = '/pinger/check',
    DeleteAccessLog = '/access/log',
    GetAccessLogPage = '/access/log/page',
    GetAccessLogGroupPage = '/access/log/group/page',
    GetAccessLogDetilPage = '/access/log/detil/page',
    PostApi = '/api',
    DeleteApi = '/api',
    GetApi = '/api',
    PutApi = '/api',
    GetBookingPage = '/booking/page',
    DeleteExceptionLog = '/exception/log',
    GetExceptionLogPage = '/exception/log/page',
    GetProxy = '/proxy',
    PostProxy = '/proxy',
    PutProxy = '/proxy',
    DeleteProxy = '/proxy',
    GetProxy1 = '/proxy/1',
    GetProxyContent = '/proxy/content',
    PostProxyContent = '/proxy/content',
    PutProxyContent = '/proxy/content',
    DeleteProxyContent = '/proxy/content',
    GetProxy2 = '/proxy/2',
    GetPublicInit = '/public/init',
    PutPublicRefresh = '/public/refresh',
    PostPublicWow = '/public/wow',
    PostPublicLogin = '/public/login',
    PostPublicTestLogin = '/public/test/login',
    DeletePublicLogout = '/public/logout',
    GetRestaurantPage = '/restaurant/page',
    GetRestaurantCombobox = '/restaurant/combobox',
    PostRole = '/role',
    GetRolePage = '/role/page',
    PutRole = '/role',
    GetRoleCombobox = '/role/combobox',
    GetRolePermission = '/role/permission',
    PutRolePermission = '/role/permission',
    PostRoute = '/route',
    DeleteRoute = '/route',
    GetRoute = '/route',
    PutRoute = '/route',
    GetRouteFolder = '/route/folder',
    GetSelf = '/self',
    PutSelf = '/self',
    PutSelfPassword = '/self/password',
    PutSelfPasswordReset = '/self/password/reset',
    DeleteSession = '/session',
    GetSession = '/session',
    GetSessionGroup = '/session/group',
    PostSubscription = '/subscription',
    DeleteSubscription = '/subscription',
    GetSubscriptionSelf = '/subscription/self',
    PutUserDisable = '/user/disable',
    PutUserEnable = '/user/enable',
    PostUser = '/user',
    PutUserLock = '/user/lock',
    PutUserUnlock = '/user/unlock',
    GetUserPage = '/user/page',
    PutUser = '/user',
    PutUserPassword = '/user/password',
    GetUserCombobox = '/user/combobox',
    PutUserPasswordReset = '/user/password/reset',
}